requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.42.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
split-folders==0.5.1
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.157
urllib3==2.0.4
Werkzeug==2.3.7
